third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/remove-styles-in-shadow-crash-2-expected.txt

CONSOLE ERROR: line 13: Uncaught TypeError: host.createShadowRoot is not a function
